+++ src/components/meal-card/EditableMealCard.tsx
@@ -1,10 +1,10 @@
 import _ from "lodash";
 import { Card, ListGroup } from "react-bootstrap";
+import { MealState } from "../../features/day-page/mealStatesSlice";
 import AddFoodInputForm from "../../features/input-form/AddFoodInputForm";
 import UpdateFoodInputForm from "../../features/input-form/UpdateFoodInputForm";
 import EditFoodButton from "../../features/meal-card/EditFoodButton";
 import NewFoodButton from "../../features/meal-card/NewFoodButton";
-import { MealState } from "../../model/MealState";
 import { Meal } from "../../model/Food";
 import { FoodItem } from "./FoodItem";
 import { AddMealCardHeader, DefaultMealCardHeader, EditMealCardHeader } from "./MealCardHeader";
